Add files via upload
UIPath Automation
The process to be automated through the use of an RPA is called "Cryptocurrency Risk Report". Cryptocurrency Risk Report".
On a daily basis, the company secretary must inform by e-mail to his boss about the risk of buying certain cryptocurrencies. his boss about the risk of buying certain cryptocurrencies.
To do so, the secretary enters a website from where he/she extracts all the information in a table format. in table format.
Once the data has been copied from the website, the secretary pastes the data into excel and evaluates, based on the variation of the last 7 days, whether or not it is risky to invest in each cryptocurrency. cryptocurrency. The risk level for each cryptocurrency will be low if the percentage variation is greater than 0". percentage variation is greater than 0", or a high risk if the variation is less than 0. The final conclusion will be added in the column "Investment_Advice (Risk Level)".
Once the excel with all the relevant information is ready, it is transformed into PDF and then attached in an email that the secretary must send every day to his boss, thus ending the process.